// libs/features/benefits/src/lib/shared/models/benefit-overview-details.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BenefitOverview } from './benefit-overview';

export class BenefitOverviewDetails {
  annuities: BenefitOverview;
  occ: BenefitOverview;
  ui: BenefitOverview;
}